+++ domain/entities/__init__.py
@@ -2,5 +2,5 @@ from domain.entities.entity import Entity
 from domain.entities.libraries import Library
 from domain.entities.loan import Loan
 from domain.entities.people import Person, PersonBorrower
-from domain.entities.people.borrower import Borrower
+from domain.entities.borrower import Borrower
 from domain.entities.thing import Thing
