+++ domain/entities/libraries/__init__.py
@@ -1,5 +1,4 @@
 from domain.entities.libraries.library import Library
 from domain.entities.libraries.library_fee import LibraryFee
-from domain.entities.libraries.base_library import BaseLibrary
 from domain.entities.libraries.simple_library import SimpleLibrary
 from domain.entities.libraries.distributed_library import DistributedLibrary
