+++ domain/entities/libraries/library_fee.py
@@ -1,10 +1,11 @@
 from domain.entities.entity import Entity
-from domain.value_items import ID
+from domain.value_items import ID, Money
 
 
 class LibraryFee(Entity):
     library_fee_id: ID
     library_id: ID
+    amount: Money
 
     @property
     def entity_id(self) -> ID:
