+++ domain/entities/people/__init__.py
@@ -1,4 +1,3 @@
-from domain.entities.people.borrower import Borrower
 from domain.entities.people.person import Person
 from domain.entities.people.person_borrower import PersonBorrower
 from domain.entities.people.user import User
