--- domain/entities/people/borrower.py
@@ -1,20 +1,0 @@
-from abc import abstractmethod
-from typing import Self, Sequence
-
-from domain.entities.entity import Entity
-from domain.entities.libraries.library_fee import LibraryFee
-from domain.value_items import ID, BorrowerVerificationFlags
-
-
-class Borrower(Entity):
-    library_id: ID  # the library this borrower is a member of
-    verification_flags: list[BorrowerVerificationFlags]
-
-    @property
-    @abstractmethod
-    def fees(self) -> Sequence[LibraryFee]:
-        raise NotImplementedError
-
-    @abstractmethod
-    def apply_fee(self, fee: LibraryFee) -> Self:
-        raise NotImplementedError
